Add Malawian Kwacha as currency
Since currency is currently not configurable in the backend, the decision was made to change it in the UI config file.
LoginService used to call backend for currency settings and if that API call failed it then used config file.
Now, in the mw-reference-ui LoginService always takes currency settings from the config file.